requirements.txt
google-adk[toolbox]
google-genai
python-dotenv
matplotlib
